Don’t print warnings when reaching dead ends
Previous behavior has been to simply iterate over and over on a dead-end node until the damping factor causes a jump to a new random node. This seems wildly inefficient, and prints a lot of noise to the log. Instead, log it as a debug and choose a random next node (as the log message actually says.)